requirements.txt
llama-index
llama-index-core
python-dotenv
llama-index-llms-groq
llama-index-embeddings-huggingface
llama-index-embeddings-google-genai
chromadb
llama-index-vector-stores-chroma
llama-parse
streamlit